+++ src/app/run-page/components/run-log/log-search/log-search.component.html
@@ -62,7 +62,9 @@
                     <p>Search run log results: <strong>{{page}} of {{pages}}</strong></p>
                 </div>
             </div>
-            <div ibmRow class="rw r6"></div>
+            <div ibmRow class="rw r6">
+                <app-log [runlog]="runlog"></app-log>
+            </div>
         </div>
         <div ibmCol class="col2" [columnNumbers]="{'md': 3, 'lg': 3}">
             <app-log-scroller></app-log-scroller>
