+++ src/app/run-page/components/run-log/log-search/log-search.component.scss
@@ -143,3 +143,11 @@ p {
 .time-elem{
   display:inline-flex;
 }
+
+.r6{
+  overflow:scroll;
+}
+
+#runlog{
+  font-size: 0.7em;
+}
