+++ src/app/run-page/components/run-log/run-log.component.html
@@ -1,11 +1,10 @@
 <div class="tabs" >
 <ibm-tabs [followFocus]=true [cacheActive]=true >
-
     <ibm-tab class=" tab bx--tabs__nav-link" id="search" heading="Search">
-        <app-log-search class="run-search" [testStructure]="testStructure"></app-log-search>
+        <app-log-search class="run-search" [testStructure]="testStructure" [runlog]="runlog"></app-log-search>
     </ibm-tab>
     <ibm-tab class="tab bx--tabs__nav-link" id="regex" heading="Regex">     
-        <app-log-regex class="run-regex" ></app-log-regex>
+        <app-log-regex class="run-regex" [runlog]="runlog"></app-log-regex>
     </ibm-tab>
 </ibm-tabs>
 </div>
