+++ src/app/run-page/components/run-log/run-log.component.ts
@@ -1,5 +1,9 @@
 import { Component, Input, OnInit, SimpleChanges } from '@angular/core';
 import { TestStructure } from 'galasa-ras-api-ts-rxjs';
+import { RasApisService } from '../../../core/rasapis.service';
+import { BootstrapService } from '../../../../app/core/bootstrap.service';
+import { HttpClient } from '@angular/common/http';
+import { ActivatedRoute } from '@angular/router';
 
 @Component({
   selector: 'app-run-log',
@@ -9,16 +13,38 @@ import { TestStructure } from 'galasa-ras-api-ts-rxjs';
 export class RunLogComponent implements OnInit {
   @Input() testStructure: TestStructure = {};
 
-  constructor() { }
+  runlog: string;
+  id: string;
+
+  constructor(private bootstrapService: BootstrapService, private http: HttpClient, private route: ActivatedRoute) {}
 
   ngOnInit(): void {
+
+    var idSub = this.route.params.subscribe(params => {
+      this.id = params['id'];
+    });
     
+    idSub.unsubscribe();
+
+    this.runlog = localStorage.getItem(this.id);
+
   }
   
-
   ngOnChanges(changes: SimpleChanges) {
-    console.log(this.testStructure.runName);
-    
+    if(this.runlog == null){
+      this.getLog(this.id);
+    }
+  }
+
+  getLog(id: string){
+    this.bootstrapService.getRasBase().then(
+      rasBase=>{
+        var url: string = rasBase.toString();
+        this.http.get(url+'/ras/run/'+`${id}`+'/runlog', {responseType:'text'}).subscribe(result=>{
+          localStorage.setItem(this.id, result);
+          this.runlog = localStorage.getItem(this.id);
+        });
+      });
   }
