+++ src/assets/apibase.js
@@ -1,3 +1,3 @@
 
-var galasaBootstrap = 'http://localhost:8080/bootstrap';
-// var galasaBootstrap = 'http://cicsk8sm.hursley.ibm.com:31410/bootstrap';
+// var galasaBootstrap = 'http://localhost:8080/bootstrap';
+var galasaBootstrap = 'http://cicsk8sm.hursley.ibm.com:31410/bootstrap';
